aas_edge_client/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.5
django-celery-beat==2.5.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
kombu==5.3.2
multidict==6.0.4
packaging==23.2
paho-mqtt==1.6.1
prompt-toolkit==3.0.39
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.7
whitenoise==6.6.0
yarl==1.9.3
